// types/Headline.ts
export type Headline = {
    id: number
    created_at: string
    country: string
    lang: string
    source: string
    section: string
    headline: string
    link: string
    emos: string
    img_src?: string
    img_alt?: string
    author?: string
    published?: string
    media_type: string
  }